+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- The option to leave, accept, or decline a game.
+
 ### Changed
 
 - hnefatafl-server-full: Reduce the information logged when run under systemd.